+++ spec/unit/infra/root/variables.tf
@@ -4,7 +4,7 @@ variable "component" {}
 variable "deployment_identifier" {}
 
 variable "subnet_ids" {
-  type = list(string)
+  type    = list(string)
   default = null
 }
 
@@ -13,7 +13,7 @@ variable "service_image" {
   default = null
 }
 variable "service_command" {
-  type = list(string)
+  type    = list(string)
   default = null
 }
 variable "service_task_container_definitions" {
@@ -30,15 +30,19 @@ variable "service_port" {
   default = null
 }
 variable "service_desired_count" {
+  type    = number
   default = null
 }
 variable "service_deployment_maximum_percent" {
+  type    = number
   default = null
 }
 variable "service_deployment_minimum_healthy_percent" {
+  type    = number
   default = null
 }
 variable "service_health_check_grace_period_seconds" {
+  type    = number
   default = null
 }
 
@@ -46,7 +50,7 @@ variable "service_role" {
   default = null
 }
 variable "service_volumes" {
-  type = list(map(string))
+  type    = list(map(string))
   default = null
 }
 
@@ -55,11 +59,12 @@ variable "scheduling_strategy" {
 }
 
 variable "placement_constraints" {
-  type = list(map(string))
+  type    = list(map(string))
   default = null
 }
 
 variable "attach_to_load_balancer" {
+  type    = bool
   default = null
 }
 variable "service_elb_name" {
@@ -76,9 +81,11 @@ variable "target_port" {
 }
 
 variable "register_in_service_discovery" {
+  type    = bool
   default = null
 }
 variable "service_discovery_create_registry" {
+  type    = bool
   default = null
 }
 variable "service_discovery_namespace_id" {
@@ -98,32 +105,40 @@ variable "service_discovery_container_port" {
 }
 
 variable "associate_default_security_group" {
+  type    = bool
   default = null
 }
 variable "include_default_ingress_rule" {
+  type    = bool
   default = null
 }
 variable "include_default_egress_rule" {
+  type    = bool
   default = null
 }
 variable "default_security_group_ingress_cidrs" {
-  type = list(string)
+  type    = list(string)
   default = null
 }
 variable "default_security_group_egress_cidrs" {
-  type = list(string)
+  type    = list(string)
   default = null
 }
 
 variable "include_log_group" {
+  type    = bool
   default = null
 }
 variable "log_group_retention" {
+  type    = number
   default = null
 }
+
 variable "force_new_deployment" {
+  type    = bool
   default = null
 }
 variable "wait_for_steady_state" {
+  type    = bool
   default = null
 }